Move Animation, GameObject, Light, MaterialBase, MatricesCb, Mesh, ModelLoader to Framework
